requirements.txt
asgiref==3.5.2
attrs==22.1.0
click==8.1.3
Django==3.2.14
django-debug-toolbar==3.5.0
django-extensions==3.2.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.23.1
drf-spectacular-sidecar==2022.8.1
gunicorn==20.1.0
h11==0.13.0
inflection==0.5.1
jsonschema==4.9.1
pendulum==2.1.2
Pillow==9.2.0
PyJWT==2.4.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
pytzdata==2020.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
uvicorn==0.18.2
